+++ plib/resources/locales/en-US.php
@@ -14,12 +14,12 @@ $messages = array(
     'removeToolTitle' => 'Remove',
     'removeToolDescription' => 'Remove the selected slave configurations.',
     'resyncToolTitle' => 'Resync',
-    'resyncToolDescription' => 'Show Plesk command for resync all slaves with master.',
+    'resyncToolDescription' => 'Resync all DNS zones from master to slaves.',
 
     'addPageTitle' => 'Creating Remote Slave Configuration',
     'howTo' => 'Add to named.conf on the remote server (config location depends on the OS of the slave server)',
     'learnMore' => 'Learn more about the slave server setup.',
-    'pleskIpLabel' => 'Plesk IP address',
+    'masterIpLabel' => 'Plesk (master) IP address',
     'ipLabel' => 'Slave DNS IP address',
     'invalidIpExistingSlave' => 'This Slave DNS already exist',
     'portLabel' => 'Port',
@@ -33,11 +33,11 @@ $messages = array(
     'viewPageTitle' => 'Remote Slave Configuration',
     'pathLabel' => 'Config path',
 
-    'resyncPageTitle' => 'Resync slaves DNS',
-
     'emptySelection' => 'Select items first.',
     'removeConfirmation' => 'Remove the selected configurations?',
-    'removeConfirmationYes' => 'Yes',
-    'removeConfirmationNo' => 'No',
+    'resyncConfirmation' => 'Resync all DNS zones from master to slaves?',
+    'confirmationYes' => 'Yes',
+    'confirmationNo' => 'No',
+    'resyncDone' => 'All DNS zones were synced from master to slaves.',
     'slaveRemoved' => 'The slave configurations were removed.',
 );